requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.1
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
colorama==0.4.6
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
playwright==1.57.0
playwright-stealth==2.0.1
propcache==0.4.1
pyee==13.0.0
python-dotenv==1.2.1
selectolax==0.4.6
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
win32_setctime==1.2.0
yarl==1.22.0
